+++ tests/conftest.py
@@ -1,0 +1,46 @@
+import pytest
+from app.mongo_db import users_collection
+from datetime import datetime
+
+@pytest.fixture(scope="session", autouse=True)
+def manage_user():
+  def create_test_user():
+    new_user =  {
+        "first_name": "Test",
+        "last_name": "Test",
+        "created_at": datetime.now(),
+        "password": "$2b$12$c/hk9viBU9LLX1I2FRcYGuijgxv6Js0gf3vV0rLfsiNkwJ/CmGeR.",
+        "email": "test@gmail.com",
+    }
+    result = users_collection.insert_one(new_user)
+    created_user = users_collection.find_one({"_id": result.inserted_id})
+    return created_user
+  
+  user = create_test_user()
+  user_id = user['_id']
+
+  yield 
+
+  def delete_test_user():
+    users_collection.delete_one({"_id": user_id})
+  
+  delete_test_user()
+
+@pytest.fixture()
+def test_user_id():
+  user = users_collection.find({"first_name": "Test",
+        "last_name": "Test","password": "$2b$12$c/hk9viBU9LLX1I2FRcYGuijgxv6Js0gf3vV0rLfsiNkwJ/CmGeR.",
+        "email": "test@gmail.com",})
+  user_id = user[-1]['_id']
+  return user_id
+
+
+
+# @pytest.fixture(autouse=True)
+# def get_auth_header(client):
+#     login_response = client.post(
+#         "/login", data={"username": "Test", "password": "1234"}
+#     )
+#     token = login_response.json().get("access_token")
+#     auth_header = {"Authorization": f"Bearer {token}"}
+#     return auth_header
